Add .mvn alongside gemspec_pom.rb
Maven was fixed in https://github.com/apache/maven/pull/94 to only
search the target pom's directory structure for .mvn when passing
-f to indicate a specific pom file. Unfortunately that breaks
logic in ruby-maven intended to temporarily copy the
.mvn/extensions.xml file to the current directory to enable
Polyglot Ruby poms. This in turn breaks jar-dependencies when
running with newer ruby-maven-libs at the root of the filesystem,
for reasons described in jruby/jruby#7059:

https://github.com/jruby/jruby/issues/7059#issuecomment-2190953877

This is a possible workaround, by putting the extension config
in the same dir as the files we want to treat as Polyglot Ruby pom
scripts. It does not fix the underlying problem with ruby-maven.